Release 3.1.0 (The Secret of Monkey Island)
Updating the Antora UI to work with Antora 3.1. The Antora Lunr extension got upgraded and enabled the Lunr extension
feature to limit search on project scope by default with a check box.

Breaking: The search highlight had to be removed and need to be reworked in upstream Antora Lunr extension instead of
a local fork.